move purge_data.py and accessdb.py inside pybossa folder
